Create own class for no history task up to date state
