+++ subprojects/core/src/main/java/org/gradle/api/internal/changedetection/changes/DefaultTaskArtifactStateRepository.java
@@ -28,6 +28,8 @@ import org.gradle.api.internal.TaskExecutionHistory;
 import org.gradle.api.internal.TaskInternal;
 import org.gradle.api.internal.changedetection.TaskArtifactState;
 import org.gradle.api.internal.changedetection.TaskArtifactStateRepository;
+import org.gradle.api.internal.changedetection.rules.DefaultTaskUpToDateState;
+import org.gradle.api.internal.changedetection.rules.NoHistoryTaskUpToDateState;
 import org.gradle.api.internal.changedetection.rules.TaskStateChange;
 import org.gradle.api.internal.changedetection.rules.TaskUpToDateState;
 import org.gradle.api.internal.changedetection.state.CurrentTaskExecution;
@@ -199,9 +201,13 @@ public class DefaultTaskArtifactStateRepository implements TaskArtifactStateRepo
         private TaskUpToDateState getStates() {
             if (states == null) {
                 HistoricalTaskExecution previousExecution = history.getPreviousExecution();
-                // Calculate initial state - note this is potentially expensive
-                CurrentTaskExecution currentExecution = history.getCurrentExecution();
-                states = new TaskUpToDateState(previousExecution, currentExecution, task);
+                if (previousExecution == null) {
+                    states = NoHistoryTaskUpToDateState.INSTANCE;
+                } else {
+                    // Calculate initial state - note this is potentially expensive
+                    CurrentTaskExecution currentExecution = history.getCurrentExecution();
+                    states = new DefaultTaskUpToDateState(previousExecution, currentExecution, task);
+                }
             }
             return states;
         }
