+++ subprojects/core/src/main/java/org/gradle/api/internal/changedetection/rules/TaskUpToDateState.java
@@ -16,91 +16,33 @@
 
 package org.gradle.api.internal.changedetection.rules;
 
-import org.gradle.api.internal.TaskInternal;
-import org.gradle.api.internal.changedetection.state.TaskExecution;
+import org.gradle.api.NonNullApi;
 
 /**
  * Represents the complete changes in a tasks state
  */
-public class TaskUpToDateState {
+@NonNullApi
+public interface TaskUpToDateState {
 
-    public static final int MAX_OUT_OF_DATE_MESSAGES = 3;
-
-    private final TaskStateChanges inputFileChanges;
-    private final OutputFileTaskStateChanges outputFileChanges;
-    private final TaskStateChanges allTaskChanges;
-    private final TaskStateChanges rebuildChanges;
-    private final TaskStateChanges outputFilePropertyChanges;
-    private final boolean hasLastExecution;
-
-    public TaskUpToDateState(TaskExecution lastExecution, TaskExecution thisExecution, TaskInternal task) {
-        hasLastExecution = lastExecution != null;
-        if (!hasLastExecution) {
-            NoHistoryTaskStateChanges noHistoryState = new NoHistoryTaskStateChanges();
-            this.allTaskChanges = new SummaryTaskStateChanges(MAX_OUT_OF_DATE_MESSAGES, noHistoryState);
-            this.rebuildChanges = new SummaryTaskStateChanges(1, noHistoryState);
-            this.inputFileChanges = null;
-            this.outputFileChanges = null;
-            this.outputFilePropertyChanges = null;
-        } else {
-            TaskStateChanges previousSuccessState = new PreviousSuccessTaskStateChanges(lastExecution);
-            TaskStateChanges taskTypeState = new TaskTypeTaskStateChanges(lastExecution, thisExecution, task);
-            TaskStateChanges inputPropertyChanges = new InputPropertyTaskStateChanges(lastExecution, thisExecution, task);
-            TaskStateChanges inputPropertyValueChanges = new InputPropertyValueTaskStateChanges(lastExecution, thisExecution, task);
-
-            // Capture outputs state
-            this.outputFilePropertyChanges = new OutputFilePropertyTaskChanges(lastExecution, thisExecution, task);
-            OutputFileTaskStateChanges uncachedOutputChanges = new OutputFileTaskStateChanges(lastExecution, thisExecution);
-            TaskStateChanges outputFileChanges = caching(uncachedOutputChanges);
-            this.outputFileChanges = uncachedOutputChanges;
-
-            // Capture input files state
-            TaskStateChanges inputFilePropertyChanges = new InputFilePropertyTaskStateChanges(lastExecution, thisExecution, task);
-            TaskStateChanges directInputFileChanges = new InputFileTaskStateChanges(lastExecution, thisExecution);
-            TaskStateChanges inputFileChanges = caching(directInputFileChanges);
-            this.inputFileChanges = new ErrorHandlingTaskStateChanges(task, inputFileChanges);
-
-            // Capture discovered inputs state from previous execution
-            DiscoveredInputTaskStateChanges discoveredChanges = new DiscoveredInputTaskStateChanges(lastExecution, thisExecution);
-            TaskStateChanges discoveredInputFileChanges = caching(discoveredChanges);
-
-            this.allTaskChanges = new ErrorHandlingTaskStateChanges(task, new SummaryTaskStateChanges(MAX_OUT_OF_DATE_MESSAGES, previousSuccessState, taskTypeState, inputPropertyChanges, inputPropertyValueChanges, outputFilePropertyChanges, outputFileChanges, inputFilePropertyChanges, inputFileChanges, discoveredInputFileChanges));
-            this.rebuildChanges = new ErrorHandlingTaskStateChanges(task, new SummaryTaskStateChanges(1, previousSuccessState, taskTypeState, inputPropertyChanges, inputPropertyValueChanges, inputFilePropertyChanges, outputFilePropertyChanges, outputFileChanges));
-        }
-    }
-
-    private static TaskStateChanges caching(TaskStateChanges wrapped) {
-        return new CachingTaskStateChanges(MAX_OUT_OF_DATE_MESSAGES, wrapped);
-    }
+    int MAX_OUT_OF_DATE_MESSAGES = 3;
 
     /**
      * Returns changes to input files only.
      */
-    public TaskStateChanges getInputFilesChanges() {
-        if (inputFileChanges == null) {
-            throw new IllegalStateException("Input file changes can only be queried when the task has been executed before");
-        }
-        return inputFileChanges;
-    }
+    TaskStateChanges getInputFilesChanges();
 
     /**
      * Returns if any output files have been changed, added or removed.
      */
-    public boolean hasAnyOutputFileChanges() {
-        return !hasLastExecution || outputFilePropertyChanges.iterator().hasNext() || outputFileChanges.hasAnyChanges();
-    }
+    boolean hasAnyOutputFileChanges();
 
     /**
      * Returns any change to task inputs or outputs.
      */
-    public TaskStateChanges getAllTaskChanges() {
-        return allTaskChanges;
-    }
+    TaskStateChanges getAllTaskChanges();
 
     /**
      * Returns changes that would force an incremental task to fully rebuild.
      */
-    public TaskStateChanges getRebuildChanges() {
-        return rebuildChanges;
-    }
+    TaskStateChanges getRebuildChanges();
 }
